refactor: :construction: Move events processing for forest root changing into MSP/BSP specific handlers
